lib/ukmpi: Provide isr-safe mbox routines
