+++ lib/ukmpi/exportsyms.uk
@@ -2,7 +2,9 @@ uk_mbox_create
 uk_mbox_free
 uk_mbox_post
 uk_mbox_post_try
+uk_mbox_post_try_isr
 uk_mbox_post_to
 uk_mbox_recv
 uk_mbox_recv_try
+uk_mbox_recv_try_isr
 uk_mbox_recv_to
