+++ lib/ukmpi/mbox.c
@@ -1,16 +1,7 @@
 #include <uk/mbox.h>
 #include <uk/assert.h>
 #include <uk/arch/limits.h>
-
-struct uk_mbox {
-	size_t len;
-	struct uk_semaphore readsem;
-	long readpos;
-	struct uk_semaphore writesem;
-	long writepos;
-
-	void *msgs[];
-};
+#include "mbox_defs.h"
 
 struct uk_mbox *uk_mbox_create(struct uk_alloc *a, size_t size)
 {
@@ -49,29 +40,6 @@ void uk_mbox_free(struct uk_alloc *a, struct uk_mbox *m)
 	uk_free(a, m);
 }
 
-/* Posts the "msg" to the mailbox, internal version that actually does the
- * post.
- */
-static inline void _do_mbox_post(struct uk_mbox *m, void *msg)
-{
-	/* The caller got a semaphore token, so we are now allowed to increment
-	 * writer, but we still need to prevent concurrency between writers
-	 * (interrupt handler vs main)
-	 */
-	unsigned long irqf;
-
-	UK_ASSERT(m);
-
-	irqf = ukplat_lcpu_save_irqf();
-	m->msgs[m->writepos] = msg;
-	m->writepos = (m->writepos + 1) % m->len;
-	UK_ASSERT(m->readpos != m->writepos);
-	ukplat_lcpu_restore_irqf(irqf);
-	uk_pr_debug("Posted message %p to mailbox %p\n", msg, m);
-
-	uk_semaphore_up(&m->readsem);
-}
-
 void uk_mbox_post(struct uk_mbox *m, void *msg)
 {
 	/* The caller got a semaphore token, so we are now allowed to increment
@@ -103,27 +71,6 @@ __nsec uk_mbox_post_to(struct uk_mbox *m, void *msg, __nsec timeout)
 	ret = uk_semaphore_down_to(&m->writesem, timeout);
 	if (ret != __NSEC_MAX)
 		_do_mbox_post(m, msg);
-	return ret;
-}
-
-/*
- * Fetch a message from a mailbox. Internal version that actually does the
- * fetch.
- */
-static inline void *_do_mbox_recv(struct uk_mbox *m)
-{
-	unsigned long irqf;
-	void *ret;
-
-	uk_pr_debug("Receive message from mailbox %p\n", m);
-	irqf = ukplat_lcpu_save_irqf();
-	UK_ASSERT(m->readpos != m->writepos);
-	ret = m->msgs[m->readpos];
-	m->readpos = (m->readpos + 1) % m->len;
-	ukplat_lcpu_restore_irqf(irqf);
-
-	uk_semaphore_up(&m->writesem);
-
 	return ret;
 }
